+++ src/main/java/com/seriesrecap/SeriesRecap/Repositories/RatesRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RatesRepository extends JpaRepository <Rates, Long> {
+public interface RatesRepository extends JpaRepository <Rates, Integer> {
 }
